Rename figure_extractor module to pdf
The module now handles PDF downloading, text extraction, and figure
extraction — "figure_extractor" no longer describes its scope. Update
all imports, tests, and CLAUDE.md references.
